+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## vNEXT
 
+- Implement `VoucherHub` upgrade and test it on GitHub Action CI using a local fork. (#57)
 - Relocate `decimals()` function in VoucherHub smart contract. (#56)
 - Better handling for env variables. (#55)
 - Remove references to blockscout v5. (#49)